[LARAVEL] - Cập nhật migrate patients

+++ database/migrations/2024_09_23_132129_create_patients_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->boolean('gender');
             $table->date('birthday');
             $table->string('address', 255);
-            $table->integer('insurance_number')->nullable(); 
-            $table->integer('emergency_contact')->nullable();
+            $table->bigInteger('insurance_number')->nullable(); 
+            $table->bigInteger('emergency_contact')->nullable();
             $table->string('occupation', 30)->nullable();
             $table->string('national');
             $table->string('phone',10)->nullable();
